Separated the parser and query builder code to remove duplicated code (also added date fields for printings)
