+++ lib/printing_search_query_builder.rb
@@ -1,300 +1,56 @@
-class PrintingSearchQueryBuilder
-    @@parser = PrintingSearchParser.new
-    @@array_keywords = [
-        'card_pool',
-        'card_subtype',
-        'card_subtype_id',
-        'eternal_points',
-        'format',
-        'has_global_penalty',
-        'illustrator_id',
-        'is_banned',
-        'is_restricted',
-        'restriction_id',
-        's',
-        'snapshot',
-        'universal_faction_cost',
-    ]
-    @@boolean_keywords = [
-        'additional_cost',
-        'advanceable',
-        'b',
-        'banlist',
-        'gains_subroutines',
-        'in_restriction',
-        'interrupt',
-        'is_unique',
-        'on_encounter_effect',
-        'performs_trace',
-        'trash_ability',
-        'u',
-    ]
-    @@date_keywords = [
-        'r',
-        'release_date'
-    ]
-    @@numeric_keywords = [
-        'advancement_cost',
-        'agenda_points',
-        'base_link',
-        'cost',
-        'g',
-        'h',
-        'influence_cost',
-        'l',
-        'link_provided',
-        'm',
-        'memory_usage',
-        'mu_provided',
-        'n',
-        'num_printed_subroutines',
-        'num_printings',
-        'o',
-        'p',
-        'quantity',
-        'recurring_credits_provided',
-        'strength',
-        'trash_cost',
-        'v',
-        'y',
-    ]
-    @@string_keywords = [
-        '_',
-        'attribution',
-        'card_type',
-        'd',
-        'f',
-        'faction',
-        'flavor',
-        'i',
-        'illustrator',
-        'r',
-        'release_date',
-        'side',
-        't',
-        'text',
-        'title',
-        'x',
-    ]
-    @@array_operators = {
-        ':' => '',
-        '!' => 'NOT',
-    }
-    @@boolean_operators = {
-        ':' => '=',
-        '!' => '!=',
-    }
-    @@date_operators = {
-        ':' => '=',
-        '!' => '!=',
-        '<' => '<',
-        '<=' => '<=',
-        '>' => '>',
-        '>=' => '>='
-    }
-    @@numeric_operators = {
-        ':' => '=',
-        '!' => '!=',
-        '<' => '<',
-        '<=' => '<=',
-        '>' => '>',
-        '>=' => '>='
-    }
-    @@string_operators = {
-        ':' => 'LIKE',
-        '!' => 'NOT LIKE',
-    }
-    @@term_to_field_map = {
-        '_' => 'unified_printings.stripped_title',
-        'a' => 'unified_printings.flavor',
-        'additional_cost' => 'unified_printings.additional_cost',
-        'advanceable' => 'unified_printings.advanceable',
-        'advancement_cost' => 'unified_printings.advancement_requirement',
-        'agenda_points' => 'unified_printings.agenda_points',
-        'attribution' => 'unified_printings.attribution',
-        'base_link' => 'unified_printings.base_link',
-        'c' => 'unified_printings.card_cycle_id',
-        'card_cycle' => 'unified_printings.card_cycle_id',
-        'card_pool' => 'unified_printings.card_pool_ids',
-        'card_set' => 'unified_printings.card_set_id',
-        'card_subtype' => 'unified_printings.lower_card_subtype_names',
-        'card_subtype_id' => 'unified_printings.card_subtype_ids',
-        'card_type' => 'unified_printings.card_type_id',
-        'cost' => 'unified_printings.cost',
-        'd' => 'unified_printings.side_id',
-        'e' => 'unified_printings.card_set_id',
-        'eternal_points' => 'unified_printings.restrictions_points',
-        'f' => 'unified_printings.faction_id',
-        'faction' => 'unified_printings.faction_id',
-        'flavor' => 'unified_printings.flavor',
-        'format' => 'unified_printings.format_ids',
-        'g' => 'unified_printings.advancement_requirement',
-        'gains_subroutines' => 'unified_printings.gains_subroutines',
-        'has_global_penalty' => 'unified_printings.restrictions_global_penalty',
-        'h' => 'unified_printings.trash_cost',
-        'illustrator_id' => 'unified_printings.illustrator_ids',
-        'i' => 'unified_printings.display_illustrators',
-        'illustrator' => 'unified_printings.display_illustrators',
-        'in_restriction' => 'unified_printings.in_restriction',
-        'influence_cost' => 'unified_printings.influence_cost',
-        'interrupt' => 'unified_printings.interrupt',
-        'is_banned' => 'unified_printings.restrictions_banned',
-        'is_restricted' => 'unified_printings.restrictions_restricted',
-        'is_unique' => 'unified_printings.is_unique',
-        'l' => 'unified_printings.base_link',
-        'link_provided' => 'unified_printings.link_provided',
-        'm' => 'unified_printings.memory_cost',
-        'memory_usage' => 'unified_printings.memory_cost',
-        'mu_provided' => 'unified_printings.mu_provided',
-        'n' => 'unified_printings.influence_cost',
-        'num_printed_subroutines' => 'unified_printings.num_printed_subroutines',
-        'num_printings' => 'unified_printings.num_printings',
-        'o' => 'unified_printings.cost',
-        'on_encounter_effect' => 'unified_printings.on_encounter_effect',
-        'p' => 'unified_printings.strength',
-        'performs_trace' => 'unified_printings.performs_trace',
-        'quantity' => 'unified_printings.quantity',
-        'r' => 'unified_printings.date_release',
-        'recurring_credits_provided' => 'unified_printings.recurring_credits_provided',
-        'release_date' => 'unified_printings.date_release',
-        'restriction_id' => 'unified_printings.restriction_ids',
-        's' => 'unified_printings.lower_card_subtype_names',
-        'side' => 'unified_printings.card_side_id',
-        'strength' => 'unified_printings.strength',
-        't' => 'unified_printings.card_type_id',
-        'text' => 'unified_printings.stripped_text',
-        'title' => 'unified_printings.stripped_title',
-        'trash_ability' => 'unified_printings.trash_ability',
-        'trash_cost' => 'unified_printings.trash_cost',
-        'u' => 'unified_printings.is_unique',
-        'universal_faction_cost' => 'unified_printings.restrictions_universal_faction_cost',
-        'v' => 'unified_printings.agenda_points',
-        'x' => 'unified_printings.stripped_text',
-        'y' => 'unified_printings.quantity',
-    }
+require_relative 'search_query_builder'
 
-    @@term_to_left_join_map = {
-    }
+class PrintingSearchQueryBuilder < SearchQueryBuilder
 
-    def initialize(query)
-        @query = query
-        @parse_error = nil
-        @parse_tree = nil
-        @left_joins = Set.new
-        @where = ''
-        @where_values = []
-        begin
-            @parse_tree = @@parser.parse(@query)
-        rescue Parslet::ParseFailed => e
-            @parse_error = e
-        end
-        if @parse_error != nil
-            return
-        end
-        constraints = []
-        where = []
-        # TODO(plural): build in explicit support for requirements
-        #   {is_banned,is_restricted,eternal_points,has_global_penalty,universal_faction_cost} all require restriction_id, would be good to have card_pool_id as well.
-        # TODO(plural): build in explicit support for smart defaults, like restriction_id should imply is_banned = false.  card_pool_id should imply the latest restriction list.
-        @parse_tree[:fragments].each {|f|
-            if f.include?(:search_term)
-                keyword = f[:search_term][:keyword].to_s
-                match_type = f[:search_term][:match_type].to_s
-                value = f[:search_term][:value][:string].to_s.downcase
-                if @@array_keywords.include?(keyword)
-                  if @@array_operators.include?(match_type)
-                    operator = @@array_operators[match_type]
-                  else
-                    @parse_error = 'Invalid array operator "%s"' % match_type
-                    return
-                  end
-                  if value.match?(/\A(\w+)-(\d+)\Z/i)
-                    value.gsub!('-', '=')
-                  end
-                  constraints << '%s (? = ANY(%s))' % [operator, @@term_to_field_map[keyword]]
-                  where << value
-                elsif @@boolean_keywords.include?(keyword)
-                    if !['true', 'false', 't', 'f', '1', '0'].include?(value)
-                        @parse_error = 'Invalid value "%s" for boolean field "%s"' % [value, keyword]
-                        return
-                    end
-                    operator = ''
-                    if @@boolean_operators.include?(match_type)
-                        operator = @@boolean_operators[match_type]
-                    else
-                        @parse_error = 'Invalid boolean operator "%s"' % match_type
-                        return
-                    end
-                    constraints << '%s %s ?' % [@@term_to_field_map[keyword], operator]
-                    where << value
-                elsif @@date_keywords.include?(keyword)
-                    if !value.match?(/\A(\d{4}-\d{2}-\d{2}|\d{8})\Z/)
-                        @parse_error = 'Invalid value "%s" for date field "%s" - only YYYY-MM-DD or YYYYMMDD are supported.' % [value, keyword]
-                        return
-                    end
-                    operator = ''
-                    if @@date_operators.include?(match_type)
-                        operator = @@date_operators[match_type]
-                    else
-                        @parse_error = 'Invalid numeric operator "%s"' % match_type
-                        return
-                    end
-                    constraints << '%s %s ?' % [@@term_to_field_map[keyword], operator]
-                    where << value
-                elsif @@numeric_keywords.include?(keyword)
-                    if !value.match?(/\A(\d+|x)\Z/i)
-                        @parse_error = 'Invalid value "%s" for integer field "%s"' % [value, keyword]
-                        return
-                    end
-                    operator = ''
-                    if @@numeric_operators.include?(match_type)
-                        operator = @@numeric_operators[match_type]
-                    else
-                        @parse_error = 'Invalid numeric operator "%s"' % match_type
-                        return
-                    end
-                    constraints << '%s %s ?' % [@@term_to_field_map[keyword], operator]
-                    where << (value.downcase == 'x' ? -1 : value)
-                else
-                    # String fields only support : and !, resolving to to {,NOT} LIKE %value%.
-                    # TODO(plural): consider ~ for regex matches.
-                    operator = ''
-                    if @@string_operators.include?(match_type)
-                        operator = @@string_operators[match_type]
-                    else
-                        @parse_error = 'Invalid string operator "%s"' % match_type
-                        return
-                    end
-                    constraints << 'lower(%s) %s ?' % [@@term_to_field_map[keyword], operator]
-                    where << '%%%s%%' % value
-                end
-                if @@term_to_left_join_map.include?(keyword)
-                    @left_joins << @@term_to_left_join_map[keyword]
-                end
-            end
+  # TODO(plural): figure out how to do name matches that are LIKEs over elements of an array.
+  # format should implicitly use the currently active card pool and restriction lists unless another is specified.
+  @fields = [
+    FieldData.new(:array, 'unified_printings.card_cycle_ids', ['card_cycle']),
+    FieldData.new(:array, 'unified_printings.card_pool_ids', ['card_pool']),
+    FieldData.new(:array, 'unified_printings.card_set_ids', ['card_set']),
+    FieldData.new(:array, 'unified_printings.lower_card_subtype_names', ['card_subtype']),
+    FieldData.new(:array, 'unified_printings.card_subtype_ids', ['card_subtype_id']),
+    FieldData.new(:array, 'unified_printings.restrictions_points', ['eternal_points']),
+    FieldData.new(:array, 'unified_printings.format_ids', ['format']),
+    FieldData.new(:array, 'unified_printings.restrictions_global_penalty', ['has_global_penalty']),
+    FieldData.new(:array, 'unified_printings.restrictions_global_penalty', ['illustrator_id']),
+    FieldData.new(:array, 'unified_printings.restrictions_banned', ['is_banned']),
+    FieldData.new(:array, 'unified_printings.restrictions_restricted', ['is_restricted']),
+    FieldData.new(:array, 'unified_printings.restriction_ids', ['restriction_id']),
+    FieldData.new(:array, 'unified_printings.snapshot_ids', ['snapshot']),
+    FieldData.new(:array, 'unified_printings.restrictions_universal_faction_cost', ['universal_faction_cost']),
+    FieldData.new(:boolean, 'unified_printings.additional_cost', ['additional_cost']),
+    FieldData.new(:boolean, 'unified_printings.advanceable', ['advanceable']),
+    FieldData.new(:boolean, 'unified_printings.gains_subroutines', ['gains_subroutines']),
+    FieldData.new(:boolean, 'unified_printings.in_restriction', ['in_restriction']),
+    FieldData.new(:boolean, 'unified_printings.interrupt', ['interrupt']),
+    FieldData.new(:boolean, 'unified_printings.is_unique', ['is_unique', 'u']),
+    FieldData.new(:boolean, 'unified_printings.on_encounter_effect', ['on_encounter_effect']),
+    FieldData.new(:boolean, 'unified_printings.performs_trace', ['performs_trace']),
+    FieldData.new(:boolean, 'unified_printings.trash_ability', ['trash_ability']),
+    FieldData.new(:date, 'unified_printings.date_release', ['date_release', 'r']),
+    FieldData.new(:integer, 'unified_printings.advancement_requirement', ['advancement_cost', 'g']),
+    FieldData.new(:integer, 'unified_printings.agenda_points', ['agenda_points', 'v']),
+    FieldData.new(:integer, 'unified_printings.base_link', ['base_link', 'l']),
+    FieldData.new(:integer, 'unified_printings.cost', ['cost', 'o']),
+    FieldData.new(:integer, 'unified_printings.influence_cost', ['influence_cost', 'n']),
+    FieldData.new(:integer, 'unified_printings.link_provided', ['link_provided']),
+    FieldData.new(:integer, 'unified_printings.memory_cost', ['memory_usage', 'm']),
+    FieldData.new(:integer, 'unified_printings.mu_provided', ['mu_provided']),
+    FieldData.new(:integer, 'unified_printings.num_printed_subroutines', ['num_printed_subroutines']),
+    FieldData.new(:integer, 'unified_printings.num_printings', ['num_printings']),
+    FieldData.new(:integer, 'unified_printings.quantity', ['quantity', 'y']),
+    FieldData.new(:integer, 'unified_printings.recurring_credits_provided', ['recurring_credits_provided']),
+    FieldData.new(:integer, 'unified_printings.strength', ['strength', 'p']),
+    FieldData.new(:integer, 'unified_printings.trash_cost', ['trash_cost', 'h']),
+    FieldData.new(:string, 'unified_printings.attribution', ['attribution']),
+    FieldData.new(:string, 'unified_printings.card_type_id', ['card_type', 't']),
+    FieldData.new(:string, 'unified_printings.faction_id', ['faction', 'f']),
+    FieldData.new(:string, 'unified_printings.flavor', ['flavor', 'flavour', 'a']),
+    FieldData.new(:string, 'unified_printings.flavor', ['illustrator', 'i']),
+    FieldData.new(:string, 'unified_printings.side_id', ['side', 'd']),
+    FieldData.new(:string, 'unified_printings.stripped_text', ['text', 'x']),
+    FieldData.new(:string, 'unified_printings.stripped_title', ['title', '_'])
+  ]
 
-            # bare/quoted words in the query are automatically mapped to stripped_title
-            if f.include?(:string)
-                    value = f[:string].to_s.downcase
-                    operator = value.start_with?('!') ? 'NOT LIKE' : 'LIKE'
-                    value    = value.start_with?('!') ? value[1..] : value
-                    constraints << 'lower(unified_printings.stripped_title) %s ?' % operator
-                    where << '%%%s%%' % value
-            end
-        }
-        @where = constraints.join(' AND ')
-        @where_values = where
-    end
-    def parse_error
-        return @parse_error
-    end
-    def where
-        return @where
-    end
-    def where_values
-        return @where_values
-    end
-    def left_joins
-        return @left_joins.to_a
-    end
 end
